requirements.txt
customtkinter>=5.2.0
pystray>=0.19.4
Pillow>=10.0.0
pyinstaller>=6.0.0
pytest>=7.4.0
pywin32>=306; sys_platform == 'win32'
